refactor: Doctor endpoint
The Doctor endpoint has been changed so that it includes a status called verified that will be updated by an admin only. This field will always be set to false when a doctor updates his/her information so that the admin can approve the data is a valid one. The Update and Get methods have been changed accordingly