fix: Doctor login not working
The doctor couldn't login due to mismatched types of the Doctor. The Doctor Contains additional fields like 'Speicalization' and 'YearOfExperience' which are not present in the 'User' Class causing it to throw an error everytime the doctor tries to login. Now this is fixed by using a projection and brigning fields only present in the 'User' class from the database upon login