fix: Get Schedules Error
The get schedules endpoint wasn't working for both Doctors and Users even though there is an existing schedule in the database. This was, suprisingly, because of all Roles was given to every user during the login which was messing up the filtering schedule process. However now this has been fixed and a user will only be assigned a role level according to the account associated with him.