fix: Errors in the Doctor Service
The error was caused due to inconsistent use of generic methods. This was resolved by using a concrete implementation of methods in the services which make use of the generic methods defined in the UserService whenever possible.

The Change made in the previous commit was also breaking some implementation, i.e., the implementation of the interface now needed a type to be passed to it since it was generic