feat: Verify email and password reset
Implemented a verify email endpoint to allow the verification of email using some logic. The logic generates a token on the database on registration which will be sent via the email and the user can then verify the email by visiting the verify email endpoint.

The password reset and forgot password services are implemented. These work in a similar manner except that they don't have an endpoint yet.

The authservice is also refactored a bit to extract repeatedly used code into methods and adding some additional methods for the new functioanlity